Move duplicated address form code to module
This change follow a similar pattern of moving some of the duplicated code in the address lookup and manual form objects into a shared module that they then both include.
